Mise à jour du téléchargement de fichiers v-0.1.3

+++ files/uploads/delete.php
@@ -1,17 +1,25 @@
 <?php
 session_start();
 if ( isset( $_GET['id'] ) ) {
-    // var_dump( $_GET['id'] );
     $fileDir = $_GET['id'];
-    unlink( $fileDir );
-    // echo "fichier supprimé";
-    $error = 'info';
-    $message = "Fichier supprimé avec succès";
-    $_SESSION['error'] = $error;
-    $_SESSION['message'] = $message;
+   
+    if( !file_exists($fileDir) ) {
+        $_SESSION['error'] = 'warning';
+        $_SESSION['message'] = 'Ce fichier n\'existe pas';
+        header("Location: index.php");
+    }else{
+        unlink( $fileDir );
+        $error = 'info';
+        $message = "Fichier supprimé avec succès";
+        $_SESSION['error'] = $error;
+        $_SESSION['message'] = $message;
+        header("Location: index.php");
+    }
+}else{
+
+    $_SESSION['error'] = 'warning';
+    $_SESSION['message'] = 'Veuillez sélectionner un fichier';
     header("Location: index.php");
 }
-// die;
-header("Location: index.php");
 
 ?>
